+++ Cargo.toml
@@ -37,6 +37,7 @@ borsh = "1.3.1"
 bs58 = "0.5.0"
 bytes = "1.6.0"
 camino = "1.1.6"
+cfg-if = "1.0.0"
 chrono = "0.4.37"
 clap = "4.4.18"
 color-eyre = "0.6.2"
